Reformat code
